JCF: define the copy and move operations for DequeBuffer by deleting them
The deletions are performed because DequeBuffer contains member data for which
copying and moving aren't coherent concepts.

A word of warning: using gcc versions older than v9_3_0, you'll get a complaint unless
you compile with -Wno-virtual-move-assign. This is because both BufferInput and BufferOutput inherit
virtually from Buffer, and there's a risk of a double-move applying to Buffer when this is the case.