JCF: change the handle() functions of CommandFacility, ConfigurationManager and ServiceManager to return a reference
In all three classes, the handle() function is meant to provide the caller with access to an
underlying object which is owned by the class. Previously, handle() returned either a raw pointer
or a reference-to-smart-pointer; in either case, it's possible for the caller of handle to take control
of the underlying object's lifetime. With a reference, this a lot harder.