JCF: bring codebase significantly closer into compliance with the DUNE C++ Style Guide
The header mostly says it all. These are changes which can't (at
currently) get picked up by the linter. Here are some of changes made:

* Error checking
I made an effort to ensure that every function which could throw an exception is wrapped in a try-catch, and every
function which returns a status variable has that variable checked. As I made these changes under a time
constraint, where I was unable to do this, I added TODOs.

* const-ification
Many variables and functions which were logically const but not declared as such are now const.
GraphConstructor::ConstructGraph, CommandFacility::listen, etc.

* Removal of unnecessary interface
E.g., replaced non-owning pointers with references when the address of an object isn't of interest

* Removal of raw owning pointers
...and replacement with smart pointers

* Formatting of TODOs
I retroactively added names, dates and emails to TODOs which were missing these